fix: update docker file and testnode script

+++ scripts/testnode.sh
@@ -1,7 +1,5 @@
-#!/bin/bash
-
 KEY="mykey"
-CHAINID="test-1"
+CHAINID="centauri-testnet-1"
 MONIKER="localtestnet"
 KEYALGO="secp256k1"
 KEYRING="test"
@@ -10,6 +8,9 @@ LOGLEVEL="info"
 #TRACE="--trace"
 TRACE=""
 
+# validate dependencies are installed
+command -v jq > /dev/null 2>&1 || { echo >&2 "jq not installed. More info: https://stedolan.github.io/jq/download/"; exit 1; }
+
 # remove existing daemon
 rm -rf ~/.banksy*
 
@@ -17,15 +18,15 @@ centaurid config keyring-backend $KEYRING
 centaurid config chain-id $CHAINID
 
 # if $KEY exists it should be deleted
-echo "decorate bright ozone fork gallery riot bus exhaust worth way bone indoor calm squirrel merry zero scheme cotton until shop any excess stage laundry" | centaurid keys add $KEY --keyring-backend $KEYRING --algo $KEYALGO --recover
+echo "taste shoot adapt slow truly grape gift need suggest midnight burger horn whisper hat vast aspect exit scorpion jewel axis great area awful blind" | centaurid keys add $KEY --keyring-backend $KEYRING --algo $KEYALGO --recover
 
 centaurid init $MONIKER --chain-id $CHAINID 
 
-# Allocate genesis accounts (cosmos formatted addresses)
-centaurid add-genesis-account $KEY 100000000000000000000000000stake --keyring-backend $KEYRING
+# Allocate genesis accounts (centauri formatted addresses)
+centaurid add-genesis-account $KEY 10000000000000000000stake --keyring-backend $KEYRING
 
-# Sign genesis transaction
-centaurid gentx $KEY 1000000000000000000000stake --keyring-backend $KEYRING --chain-id $CHAINID
+# Sign genesis transaction centauri1594tdya20hxz7kjenkn5w09jljyvdfk8kx5rd6
+centaurid gentx $KEY 100000000000000000stake --keyring-backend $KEYRING --chain-id $CHAINID
 
 # Collect genesis tx
 centaurid collect-gentxs
@@ -39,7 +40,11 @@ fi
 
 # update request max size so that we can upload the light client
 # '' -e is a must have params on mac, if use linux please delete before run
-sed -i'' -e 's/max_body_bytes = /max_body_bytes = 1/g' ~/.banksy/config/config.toml
+sed -i'' -e 's/max_body_bytes = /max_body_bytes = 5/g' ~/.banksy/config/config.toml
+sed -i'' -e 's/rpc-max-body-bytes = /rpc-max-body-bytes = 5/g' ~/.banksy/config/app.toml
+sed -i'' -e 's/max_tx_bytes = /max_tx_bytes = 5/g' ~/.banksy/config/config.toml
+
+cat $HOME/.banksy/config/genesis.json | jq '.app_state["gov"]["params"]["voting_period"]="90s"' > $HOME/.banksy/config/tmp_genesis.json && mv $HOME/.banksy/config/tmp_genesis.json $HOME/.banksy/config/genesis.json
 
 # Start the node (remove the --pruning=nothing flag if historical queries are not needed)
-centaurid start --pruning=nothing  --minimum-gas-prices=0.0001stake --rpc.laddr tcp://0.0.0.0:26657
+centaurid start --minimum-gas-prices=0.0001stake --rpc.laddr tcp://0.0.0.0:26657
